requirements.txt
pytest
pytest-env
json
hashlib
tinydb
flask
dynaconf
pillow
pillow-heif
scikit-image
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
pickle
